+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- Return single route quote if split quote fails
 - CoinGecko pricing source support
 - Consistently rename min liqudity filter across pool models & configs to min pool liquidity capitalization (cap)
 - Validate token in and out in /router quote endpoints to ensure that they are not equal to each other & clean up tokens/prices